git-svn-id: https://svn.apache.org/repos/asf/cassandra/trunk@1081872 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/cassandra/io/util/MappedFileDataInput.java
@@ -75,7 +75,7 @@ public class MappedFileDataInput extends AbstractDataInput implements FileDataIn
         return new MappedFileDataInputMark(position);
     }
 
-    public int bytesPastMark(FileMark mark)
+    public long bytesPastMark(FileMark mark)
     {
         assert mark instanceof MappedFileDataInputMark;
         assert position >= ((MappedFileDataInputMark) mark).position;
